[feat] change landing page to be client page
