+++ components/Header.tsx
@@ -9,8 +9,8 @@ const Header: React.FC = () => {
 
   let left = (
     <div className="left">
-      <Link href="/">
-        <a className="bold" data-active={isActive("/")} id="future-health-home-page-link">
+      <Link href="/navigation">
+        <a className="bold" data-active={isActive("/navigation")} id="future-health-home-page-link">
           FutureHealth
         </a>
       </Link>
